move token info  to other constant file
