+++ src/components/swapv2/TokenInfoV2.tsx
@@ -4,7 +4,7 @@ import useTokenInfo, { TokenInfo } from 'hooks/useTokenInfo'
 import { useCurrencyConvertedToNative } from 'utils/dmm'
 import { TokenInfoWrapper } from './styleds'
 import SingleTokenInfo, { HowToSwap } from 'components/swapv2/SingleTokenInfo'
-import { TOKEN_INFO_DESCRIPTION, WHITE_LIST_TOKEN_INFO_PAIR } from 'constants/networks'
+import { TOKEN_INFO_DESCRIPTION, WHITE_LIST_TOKEN_INFO_PAIR } from 'constants/tokenLists/token-info'
 import { getNetworkSlug, getSymbolSlug } from 'utils/string'
 import { useActiveWeb3React } from 'hooks'
